Fix bug that made players lose turns for invalid move
to do this, I created an instance variable in the game class that would
serve to track the state of the program. I also modified the make_move
method to alter the state with each successful move. Then I wraped the
player-move method calls with until loops that used that state as their
condition.
now, making a wrong move will present the same player with a prompt to
make the move again, correctly this time.